Separate flows based on dependencies.
